migrate to new OAuth2 libraries
The oauth2client library has been deprecated. This change replaces its
usage with the new recommended set of libraries:

 * google-auth
 * google-auth-oauthlib
 * google-auth-httplib2

The new libraries do not provide storage capabilities similar to
oauth2client.file. Instead, we implement loading and storing the
credentials manually. The old credential storage file format is
compatible with the new libraries.

The authorize command is changed to accept a client_secrets.json file
instead of a pair client_id/client_secret. This makes it both simpler to
implement and easier to use.

The scopes have changed to the correct new scope. The new library raises
an exception if the token is granted with a differing set of scopes.